+++ student_223024/BUKEP.Student.WebFormsTodo/TaskManagerState.cs
@@ -21,7 +21,7 @@ namespace BUKEP.Student.WebFormsTodo
 					// вот тут придется сменить
 					// AttachDbFilename=E:\\workge\\student_223024\\BUKEP.Student.WebFormsTodo\\App_Data\\Database1.mdf
 					// на свою ссылку БД, не знаю как по другому
-					HttpContext.Current.Session["_userData"] = new TaskDatabase("Data Source=(LocalDB)\\MSSQLLocalDB;AttachDbFilename=E:\\workge\\student_223024\\BUKEP.Student.WebFormsTodo\\App_Data\\Database1.mdf;Integrated Security=True;");
+					HttpContext.Current.Session["_userData"] = new DatabaseTaskManager("Data Source=(LocalDB)\\MSSQLLocalDB;AttachDbFilename=E:\\workge\\student_223024\\BUKEP.Student.WebFormsTodo\\App_Data\\Database1.mdf;Integrated Security=True;");
 				}
 				return (ITaskManager)HttpContext.Current.Session["_userData"];
 			}
