+++ student_223024/BUKEP.Student.WebFormsTodo/Web.config
@@ -5,7 +5,7 @@
   -->
 <configuration>
   <connectionStrings>
-    <add name="Database" connectionString="Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=E:\workge\student_223024\BUKEP.Student.WebFormsTodo\App_Data\Database1.mdf;Integrated Security=True"></add>
+    <add name="Database" connectionString="Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\Database1.mdf;Integrated Security=True"></add>
   </connectionStrings>
   <system.web>
     <compilation debug="true" targetFramework="4.8" />
